third_party/blink/web_tests/html5lib/generated/run-tests1-write-expected.txt

CONSOLE ERROR: Uncaught SyntaxError: Unexpected token '<'
../resources/tests1.dat: PASS
